Support nested promise callbacks and fns
Lift the restriction that promise-returning callbacks and fns may only
appear at the top level. The evaluator now propagates promises at any
depth via a sequential iterator. Return type is unchanged: `boolean`
for fully synchronous expressions, `Promise<boolean>` only when an
async branch is evaluated. `$and` and `$or` still stop on the first
deciding value; spy tests verify later async items are skipped.

Other changes:
- Split the monolithic test/index.test.js into one file per category
  (primitive, callback, operator, fn, compound, shortCircuit).
- Introduce CHANGELOG.md (Keep a Changelog format) with v0.2.0 and
  v0.1.0 entries.
- Drop the stale StackOverflow credit comment.

Closes #2
